pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gbcapacity"
version = "0.1.0"
description = "Real-time UK electricity system price and NIV streaming"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "GB Capacity", email = "info@gbcapacity.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
keywords = ["electricity", "energy", "uk", "balancing", "system-price", "niv", "streaming"]
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28",
    "websocket-client>=1.5",
    "pandas>=1.5",
]

[project.urls]
Homepage = "https://gbcapacity.com"
Repository = "https://github.com/eventide99/gbcapacity"
Documentation = "https://github.com/eventide99/gbcapacity#readme"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
where = ["."]
